extract simulation mechanics for stochastic tests to separate source file
